cpp fundamentals refactored

+++ cpp/fundamentals/memory-objects/sizeof.cxx
@@ -3,8 +3,8 @@
 int main()
 {
     int shortNumber = 1000;
-    long meduimNumber = 1000000;
-    long long largeNumber = 1000000000;
+    double meduimNumber = 1000000;
+    long double largeNumber = 1000000000;
 
     // size of allocated memory depends on the variable type
     std::cout << sizeof(shortNumber) << std::endl;
